pyproject.toml
[build-system]
requires = [
    "hatchling",
    "hatch-regex-commit",
]
build-backend = "hatchling.build"

# project metadata

[project]
name = "hexdoc"
description = "Python web book docgen for Patchouli and Hex Casting."
authors = [
    { name="object-Object" },
    { name="Alwinfy" },
]
readme = "README.md"
license = { file = "LICENSE" }
keywords = [
    "hexdoc",
]
classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Java",
    "Programming Language :: JavaScript",
    "Programming Language :: Kotlin",
    "Programming Language :: Python :: 3.11",
    "Topic :: Documentation",
    "Topic :: File Formats :: JSON",
    "Topic :: Games/Entertainment",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Localization",
    "Topic :: Text Processing :: Markup :: HTML",
    "Typing :: Typed",
]
requires-python = ">=3.11"
dynamic = ["version"]
dependencies = [
    "importlib_resources~=6.0",
    "Jinja2~=3.1",
    "minecraft-render~=2.0.0a6",
    "more_itertools~=10.1",
    "ordered-set~=4.1",
    "packaging~=23.2",
    "pluggy~=1.3",
    "pydantic_settings~=2.0",
    "pydantic~=2.3,!=2.4.0",
    "pygithub~=2.1",
    "pyjson5~=1.6",
    "requests~=2.31",
    "typer[all]~=0.9",
    "typing_extensions~=4.6",

    # Favicons deps, because it's vendored
    "pillow~=9.5.0",
    "rich~=13.3.4",
    "svglib~=1.5.1",
]

[project.optional-dependencies]
pdoc = [
    "pdoc~=14.1",
    "hatch",
]
test = [
    "pyright!=1.1.335,!=1.1.336",
    "pytest~=7.4",
    "pytest-copie~=0.1",
    "pytest-dependency~=0.5",
    "syrupy~=4.6",
]
dev = [
    "hexdoc[pdoc,test]",
    "ruff==0.1.4",
    "pre-commit",
    "nox",
]

[project.urls]
Homepage = "https://book.hexxy.media"
Documentation = "https://hexdoc.hexxy.media"
Source = "https://github.com/object-Object/hexdoc"

[tool.hatch.version]
path = "src/hexdoc/__version__.py"

[tool.hatch.build]
artifacts = [
    "/src/hexdoc/_export/generated",
]

[tool.hatch.build.targets.sdist]
only-include = [
    "src",
    "vendor",
]

[tool.hatch.build.targets.wheel]
packages = [
    "src/hexdoc",
    "vendor/_hexdoc_favicons",
]

[project.scripts]
hexdoc = "hexdoc.cli.app:app"

# hexdoc

[project.entry-points.hexdoc]
hexdoc = "hexdoc._hooks:HexdocPlugin"

# Pytest

[tool.pytest.ini_options]
addopts = ["--import-mode=importlib"]
markers = ["template"]
testpaths = ["test"]
filterwarnings = [
    'ignore:setDaemon\(\) is deprecated, set the daemon attribute instead:DeprecationWarning:javascript.events',
]

[tool.coverage.report]
include_namespace_packages = true

# formatting

[tool.ruff]
extend-select = [
    "I", # import sorting
]
extend-ignore = [
    # covered by Pyright
    "F821", # undefined variable
    "F401", # imported but unused
    "F841", # unused variable
]
extend-exclude = [
    "vendor",
    "venv",
    "**/.*",
    "**/node_modules",
    "**/__pycache__",
]

[tool.ruff.lint.isort]
combine-as-imports = true

# type checking

[tool.pyright]
pythonVersion = "3.11"
pythonPlatform = "All"

venvPath = "."
venv = "venv"

include = [
    "src",
    "test",
]
extraPaths = [
    "vendor",
]
exclude = [
    "noxfile.py",
    "vendor",
    "venv",
    "**/.*",
    "**/node_modules",
    "**/__pycache__",
    "template",
]

typeCheckingMode = "basic"

enableExperimentalFeatures = true
strictDictionaryInference = true
strictListInference = true
strictSetInference = true

reportAssertAlwaysTrue = "error"
reportConstantRedefinition = "error"
reportDeprecated = "error"
reportFunctionMemberAccess = "error"
reportIncompatibleMethodOverride = "error"
reportIncompatibleVariableOverride = "error"
reportIncompleteStub = "error"
reportInconsistentConstructor = "error"
reportInvalidStringEscapeSequence = "error"
reportInvalidStubStatement = "error"
reportInvalidTypeVarUse = "error"
reportMatchNotExhaustive = "error"
reportMissingParameterType = "error"
reportOverlappingOverload = "error"
reportSelfClsParameterName = "error"
reportTypeCommentUsage = "error"
reportUnknownParameterType = "error"
reportUnnecessaryCast = "error"
reportUnnecessaryContains = "error"
reportUnsupportedDunderAll = "error"
reportUntypedBaseClass = "error"
reportUntypedClassDecorator = "error"
reportUntypedFunctionDecorator = "error"
reportUntypedNamedTuple = "error"
reportWildcardImportFromLibrary = "error"

reportMissingTypeArgument = "warning"
reportPrivateUsage = "warning"
reportUnknownArgumentType = "warning"
reportUnknownLambdaType = "warning"
reportUnknownMemberType = "warning"
reportUnknownVariableType = "warning"
reportUnnecessaryComparison = "warning"
reportUnnecessaryIsInstance = "warning"
reportUnusedClass = "warning"
reportUnusedExpression = "warning"
reportUnusedFunction = "warning"
reportUnusedImport = "warning"
reportUnusedVariable = "warning"

reportMissingTypeStubs = "none"
reportDuplicateImport = "none"
